Refactor 2   *complete food events and extract from index
