+++ lib/models/food.js
@@ -30,7 +30,7 @@ Food.delete = function(foodID) {
     }
   }).done(function(response) {
     return response
-    })
+  })
 }
 
 Food.removeFood = function(element) {
